% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_of_centre.R
\name{describe_kurtosis}
\alias{describe_kurtosis}
\title{Interpret the value given for kurtosis in 'plain' English.}
\usage{
describe_kurtosis(kurtosis)
}
\arguments{
\item{kurtosis}{value of kurtosis}
}
\value{
interpretation of given kurtosis value
}
\description{
The function takes a value of kurtosis, calculates excess kurtosis and
creates an interpretation.
}
\examples{
standard usage:
describe_kurtosis(4)
"A kurtosis of 4 results in an excess kurtosis of 1, which indicates a
leptokurtic or 'greater-tailed' distribution"

}
