% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_of_centre.R
\name{describe_skewness}
\alias{describe_skewness}
\title{Interpret the value given for skewness in 'plain' English.}
\usage{
describe_skewness(skew)
}
\arguments{
\item{skew}{value of skewness}
}
\value{
A character vector
}
\description{
Interpret the value given for skewness in 'plain' English.
}
\examples{
standard usage:
describe_skewness(1)
"1 indicates highly right skewed data."

}
