% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\docType{data}
\name{h_testing}
\alias{h_testing}
\title{List containing definitions of terms used in hypothesis testing.}
\format{
An object of class \code{list} of length 3.
}
\usage{
h_testing
}
\value{
A character vector
}
\description{
List containing definitions of terms used in hypothesis testing.
}
\examples{
standard usage:
glossary$hypothesis_testing$p_value
"The p-value is the probability that a value equal to or more extreme than the observed sample statistic would be observed if the null hypothesis were true. The smaller the calculated p-value, the more unlikely it is that the null hypothesis is true."
}
\keyword{datasets}
